agent/requirements.txt

fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.8.2
httpx==0.27.2
pyyaml==6.0.2
numpy==1.24.4
# Optional ML stack (uncomment & install on capable devices)
# torch==2.3.1
# scikit-learn==1.5.1
# flwr==1.8.0
